Refactor alert and device token management: rename classes, update package structure, and enhance data models
