+++ src/main/java/com/protectalk/alert/dto/ScamAlertReason.java
@@ -1,0 +1,5 @@
+package com.protectalk.alert.dto;
+
+public enum ScamAlertReason {
+    OK, BELOW_THRESHOLD, NO_CONTACTS, DELIVERY_FAILED, PARTIAL_DELIVERY
+}
