+++ src/main/java/com/protectalk/alert/model/RiskLevel.java
@@ -1,0 +1,7 @@
+package com.protectalk.alert.model;
+
+public enum RiskLevel {
+    GREEN,
+    YELLOW,
+    RED
+}
