--- src/main/java/com/protectalk/alerts/domain/Platform.java
@@ -1,12 +1,0 @@
-package com.protectalk.alerts.domain;
-
-public enum Platform {
-    ANDROID,
-    IOS,
-    WEB,
-    DESKTOP;
-
-    public String getValue() {
-        return name().toLowerCase();
-    }
-}
