--- src/main/java/com/protectalk/alerts/domain/RiskLevel.java
@@ -1,7 +1,0 @@
-package com.protectalk.alerts.domain;
-
-public enum RiskLevel {
-    GREEN,
-    YELLOW,
-    RED
-}
