--- src/main/java/com/protectalk/alerts/dto/ScamAlertReason.java
@@ -1,5 +1,0 @@
-package com.protectalk.alerts.dto;
-
-public enum ScamAlertReason {
-    OK, BELOW_THRESHOLD, NO_CONTACTS, DELIVERY_FAILED, PARTIAL_DELIVERY
-}
