+++ src/main/java/com/protectalk/device/model/Platform.java
@@ -1,0 +1,12 @@
+package com.protectalk.device.model;
+
+public enum Platform {
+    ANDROID,
+    IOS,
+    WEB,
+    DESKTOP;
+
+    public String getValue() {
+        return name().toLowerCase();
+    }
+}
