+++ src/main/java/com/protectalk/messaging/Delivery.java
@@ -1,0 +1,9 @@
+package com.protectalk.messaging;
+
+public record Delivery(
+    String token,
+    boolean success,
+    String messageId,
+    String errorCode,
+    String errorMessage
+) {}
