+++ src/main/java/com/protectalk/messaging/NotificationGateway.java
@@ -1,0 +1,7 @@
+package com.protectalk.messaging;
+
+import com.google.firebase.messaging.FirebaseMessagingException;
+
+public interface NotificationGateway {
+    NotificationResult send(OutboundMessage m) throws FirebaseMessagingException;
+}
