+++ src/main/java/com/protectalk/messaging/NotificationResult.java
@@ -1,0 +1,11 @@
+package com.protectalk.messaging;
+
+import java.util.List;
+
+public record NotificationResult(
+    int total,
+    int success,
+    int failure,
+    List<Delivery> deliveries,
+    List<String> invalidTokens
+) {}
