+++ src/main/java/com/protectalk/messaging/OutboundMessage.java
@@ -1,0 +1,11 @@
+package com.protectalk.messaging;
+
+import java.util.List;
+import java.util.Map;
+
+public record OutboundMessage(
+    String title,
+    String body,
+    Map<String, String> data,
+    List<String> tokens
+) {}
