+++ src/main/resources/application.properties
@@ -8,3 +8,6 @@ spring.data.mongodb.uri=mongodb://localhost:27017/protectalk
 logging.level.com.protectalk=DEBUG
 logging.level.org.springframework=INFO
 logging.level.org.mongodb.driver=INFO
+
+# Application Settings
+protectalk.threshold.modelScore=0.80
